🔨 refactor(EventPassNftFilesTable.stories.tsx): Remove unused imports and function eventPassNftFilesTableMocks 🔨 refactor(EventPassNftFilesTable/examples.tsx): Move function eventPassNftFilesTableMocks to its own file 🔨 refactor(EventSheet.stories.tsx): Remove unused imports and import eventPassNftFiles and eventPassNftFilesTableMocks directly
